+++ app/assets/javascripts/components/story/ExpandedStory/index.js
@@ -5,14 +5,15 @@ import ExpandedStoryControls from './ExpandedStoryControls';
 import ExpandedStoryEstimate from './ExpandedStoryEstimate';
 import ExpandedStoryType from './ExpandedStoryType';
 import ExpandedStoryDescription from './ExpandedStoryDescription';
-import { createTask, deleteTask, toggleTask } from '../../../actions/task';
 import ExpandedStoryNotes from './ExpandedStoryNotes';
 import ExpandedStoryState from './ExpandedStoryState';
 import ExpandedStoryTitle from './ExpandedStoryTitle';
 import ExpandedStoryLabels from './ExpandedStoryLabels';
-import { deleteNote, createNote } from '../../../actions/note';
-import { editStory, updateStory, deleteStory } from '../../../actions/story';
 import ExpandedStoryTask from './ExpandedStoryTask';
+import { editStory, updateStory, deleteStory } from '../../../actions/story';
+import { createTask, deleteTask, toggleTask } from '../../../actions/task';
+import { deleteNote, createNote } from '../../../actions/note';
+import { addLabel, removeLabel } from '../../../actions/labels';
 import { connect } from 'react-redux';
 import * as Story from '../../../models/beta/story';
 
@@ -21,12 +22,15 @@ export const ExpandedStory = ({
   onToggle,
   editStory,
   updateStory,
+  deleteStory,
   project,
   createTask,
   deleteTask,
   toggleTask,
   deleteNote,
-  createNote
+  createNote,
+  addLabel,
+  removeLabel
 }) => {
   return (
     <div className="Story Story--expanded">
@@ -59,15 +63,16 @@ export const ExpandedStory = ({
       />
 
       <ExpandedStoryLabels
-        labels={story.labels}
+        onAddLabel={(label) => addLabel(story.id, label)}
+        labels={story._editing.labels}
+        projectLabels={project.labels}
+        onRemoveLabel={(labelName) => removeLabel(story.id, labelName)}
         onEdit={(value) => editStory(story.id, { labels: value })}
       />
 
       <ExpandedStoryDescription
         story={story}
-        onToggle={(task, status) => toggleTask(project.id, story, task, status)}
-        onDelete={(taskId) => deleteTask(project.id, story.id, taskId)}
-        onSave={(task) => createTask(project.id, story.id, task)}
+        onEdit={(newAttributes) => editStory(story.id, newAttributes)}
       />
 
       <ExpandedStoryNotes
@@ -103,6 +108,8 @@ export default connect(
     toggleTask,
     deleteStory,
     deleteNote,
-    createNote
+    createNote,
+    addLabel,
+    removeLabel
   }
 )(ExpandedStory);