+++ app/assets/javascripts/models/beta/label.js
@@ -1,3 +1,5 @@
+import _ from 'underscore';
+
 export const splitLabels = (labels) => {
   if (labels) {
     return labels.split(',')
@@ -11,6 +13,18 @@ export const splitLabels = (labels) => {
 
   return [];
 };
+
+export const removeLabel = (labels, labelName) =>
+  labels.filter(label => label.name !== labelName);
+
+export const addLabel = (labels, newLabel) =>
+  uniqueLabels([
+    ...labels,
+    newLabel
+  ]);
+
+export const uniqueLabels = (labels) =>
+  _.uniq(labels, label => label.name);
 
 export const joinLabels = (labels) =>
   getNames(labels).join(',');