+++ app/assets/javascripts/models/beta/project.js
@@ -1,0 +1,14 @@
+import * as Label from './label';
+
+export const deserialize = (board) => ({
+  ...board.project,
+  labels: Label.splitLabels(board.labels)
+});
+
+export const addLabel = (project, label) => ({
+  ...project,
+  labels: Label.uniqueLabels([
+    ...project.labels,
+    label
+  ])
+});