+++ app/assets/javascripts/reducers/project.js
@@ -1,4 +1,5 @@
 import actionTypes from 'actions/actionTypes';
+import * as Project from '../models/beta/project';
 
 const initialState = {};
 
@@ -6,6 +7,8 @@ const projectReducer = (state = initialState, action) => {
   switch (action.type) {
     case actionTypes.RECEIVE_PROJECT:
       return action.data;
+    case actionTypes.ADD_LABEL_TO_PROJECT:
+      return Project.addLabel(state, action.label);
     default:
       return state;
   };