+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Notes to Story in board V2
 - Delete option to Story in board V2
 - Select to Story State in board V2
+- Labels to Story in board V2
 
 ### Fix
 - Refactor Redux flow to remove stories duplication in board V2